fix: allow passing a custom link component
