+++ src/Navigation/stories/fixtures/logos/CustomLogo3.tsx
@@ -1,6 +1,6 @@
 import React, { SVGProps } from "react";
 
-export default function Logo(props: SVGProps<SVGSVGElement>) {
+export default function CustomLogo3(props: SVGProps<SVGSVGElement>) {
   return (
     <svg viewBox="0 0 415 109" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
       <path
