--- src/Navigation/todo.md
@@ -1,6 +1,0 @@
-### What's left
-
-[ ] Cypress tests
-[ ] Fix the tooltip
-[ ] Readme.md
-[ ] Export component and types
